requirements.txt
opencv-python
numpy
pillow
streamlit
kaggle
streamlit-cropper